src/airflow/SST/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dag-generator"
version = "0.1.0"
description = "A clean and elegant Apache Airflow DAG generator"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "Jinja2>=3.0.0",
    "PyYAML>=6.0",
    "click>=8.0.0",
    "SQLAlchemy>=1.4.0",
    "pydantic>=2.0.0",
]

[project.scripts]
dag-generator = "dag_generator.cli:main"